Aligning the Fieldcollector and adding collector parsing for the DeviceRegistrationCollector
