+++ Davinci/Davinci/collector/FieldCollector.swift
@@ -2,7 +2,7 @@
 //  FieldCollector.swift
 //  PingDavinci
 //
-//  Copyright (c) 2024 - 2025 Ping Identity. All rights reserved.
+//  Copyright (c) 2024 - 2025 Ping Identity Corporation. All rights reserved.
 //
 //  This software may be modified and distributed under the terms
 //  of the MIT license. See the LICENSE file for details.
@@ -21,7 +21,9 @@ open class FieldCollector<T>: Collector, Validator, @unchecked Sendable {
     public private(set) var key: String = ""
     public private(set) var label: String = ""
     public private(set) var required: Bool = false
-    public let id = UUID()
+    public var id: String {
+        return key
+    }
   
     /// Initializes a new instance of `FieldCollector`.
     public init() {}
