+++ Davinci/Davinci/collector/LabelCollector.swift
@@ -2,7 +2,7 @@
 //  LabelCollector.swift
 //  PingDavinci
 //
-//  Copyright (c) 2025 Ping Identity. All rights reserved.
+//  Copyright (c) 2025 Ping Identity Corporation. All rights reserved.
 //
 //  This software may be modified and distributed under the terms
 //  of the MIT license. See the LICENSE file for details.
@@ -13,9 +13,11 @@ import Foundation
 
 /// Class representing a LABEL type.
 /// It conforms to the `Collector` protocol and displays a label on the form.
-public class LabelCollector: Collector, @unchecked Sendable {
+public class LabelCollector: Collector, @unchecked Sendable {    
     /// The UUID of the field collector.
-    public let id = UUID()
+    public var id: String {
+        return UUID().uuidString
+    }
     /// The label content.
     public private(set) var content: String = ""
     
@@ -23,5 +25,9 @@ public class LabelCollector: Collector, @unchecked Sendable {
     /// - Parameter json: The json to initialize from.
     public required init(with json: [String : Any]) {
         content = json[Constants.content] as? String ?? ""
+    }
+    
+    public func payload() -> Never? {
+        return nil
     }
 }
