+++ Davinci/Davinci/collector/Validator.swift
@@ -1,0 +1,14 @@
+// 
+//  Validator.swift
+//  Davinci
+//
+//  Copyright (c) 2025 Ping Identity Corporation. All rights reserved.
+//
+//  This software may be modified and distributed under the terms
+//  of the MIT license. See the LICENSE file for details.
+//
+
+
+public protocol Validator {
+    func validate() -> [ValidationError]
+}
